Handle Exceptions much better, by pushing them through the queues.
The previous handling of exceptions pushed a 'close' message through,
which fixed a deadlock-issue. The places where the queues where pop'd
checked for the wrong message, though. This fixes it, and instead pushes
the exception through, so it can be re-raised on the main thread.

This fixes the unreported demise of the server-side, and also fixes a
bug where no other exceptions got back to the client, for instance
NameError's while eval'ing code.